+++ frontend/src/components/Common/NavigationBar.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useState } from 'react';
 import styled, { css } from 'styled-components';
 import { Link, useNavigate } from 'react-router-dom';
-import searchIcon from '../assets/searchIcon.png';
+import searchIcon from '../../assets/searchIcon.png';
 
 import DropDownMenu from './DropDownMenu';
